+++ finble-backend/src/main/java/com/ssafy/BlueMarble/websocket/dto/payload/game/NtsPayload.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class NtsPayload {
     private String nickname;
-    private int taxAmount;
+    private Long taxAmount;
     private ConstructPayload.Asset updatedAsset;
 }
